issue #85: fix fast refresh warning
