+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "finesse-frontend",
   "private": true,
-  "version": "1.0.0",
+  "version": "1.0.1",
   "type": "module",
   "scripts": {
     "prestart": "ts-appversion",
