+++ src/App.tsx
@@ -2,9 +2,9 @@ import { Route, Routes } from "react-router";
 import { BrowserRouter } from "react-router-dom";
 import "./App.css";
 import { config } from "./config";
-import { AlertProvider } from "./contexts/AlertContext";
-import { DataProvider } from "./contexts/DataContext";
-import { LayoutProvider } from "./contexts/LayoutContext";
+import AlertProvider from "./contexts/AlertProvider";
+import DataProvider from "./contexts/DataProvider";
+import LayoutProvider from "./contexts/LayoutProvider";
 import Home from "./pages/home/Home";
 import SearchPage from "./pages/search/SearchPage";
