+++ src/pages/search/SearchPage.tsx
@@ -1,31 +1,36 @@
 // SearchPage.tsx
-import React, { useEffect } from "react";
-import { useApiUtil } from "../../api/useApiUtil";
+import React, { useEffect, useState } from "react";
+import { search } from "../../api/useApiUtil";
 import Header from "../../components/header/Header";
 import CFIALogo from "../../components/logo/CFIALogo";
 import SearchResultList from "../../components/search_results/SearchResultsList";
 import { SearchBar } from "../../components/searchbar/SearchBar";
-import { useData } from "../../contexts/DataContext";
+import { useData } from "../../contexts/dataContext";
 import {
   LayoutContainer,
-  SearchPageHeaderContainer,
   SearchBarSearchContainer,
+  SearchPageHeaderContainer,
 } from "../../styles/indexElements";
+import { ResponseData } from "../../types";
 
 const SearchPage: React.FC = () => {
   const { searchTerm, currentSearchSource } = useData();
   // Provide a default value for termProp
   const termProp = searchTerm !== null ? searchTerm : "";
-
-  const { data } = useApiUtil({
-    term: termProp,
-    currentSearchSource,
-  });
+  const [data, setData] = useState<ResponseData | null>(null);
 
   // useEffect hook to print the value of term whenever it changes
   useEffect(() => {
     console.log("Term:", searchTerm);
   }, [searchTerm]);
+
+  useEffect(() => {
+    search(termProp, currentSearchSource)
+      .then((data) => {
+        setData(data);
+      })
+      .catch((e) => console.error("Error using api", e));
+  }, [termProp, currentSearchSource]);
 
   return (
     <LayoutContainer>
